+++ ecs/task/variables.tf
@@ -226,7 +226,7 @@ variable "task_resource_memory_gib_default" {
 variable "task_resource_memory_host_gib_default" {
   type        = number
   default     = null
-  description = "Memory remaining for host OS. Defaults to 13/32 + Instance memory / 64"
+  description = "Memory remaining for host OS. Defaults to an empirical fit that is suitable for most applications."
 }
 
 variable "task_resource_num_vcpu_default" {