+++ iam/user/password_policy/versions.tf
@@ -1,6 +1,6 @@
 terraform {
   required_providers {
-    aws = { # tflint-ignore: terraform_unused_required_providers
+    aws = {
       source  = "hashicorp/aws"
       version = ">= 5.0"
     }