+++ module_template/variables.tf
@@ -1,6 +1,19 @@
 variable "resource_map" {
   type = map(object({
+    name_include_app_fields = optional(bool)
+    name_infix              = optional(bool)
   }))
+}
+
+variable "name_include_app_fields_default" {
+  type        = bool
+  default     = true
+  description = "If true, the Terraform project context will be included in the name"
+}
+
+variable "name_infix_default" {
+  type    = bool
+  default = true
 }
 
 variable "std_map" {